Prepared for migration to using external ini data for all settings. * Added support for fan-added theaters available in CnCNet: Snow for TD; Desert, Jungle, Barren and Cave for RA. These only work in classic mode, if their theater mix files are found in the configured classic files folder. * Mouse zoom is now disabled during drag-scroll operations, since it invariably messed up the position calculations. * Map templates, Smudge and Overlay are no longer restricted to specific theaters; if they exist in a theater, they are allowed. * Theater-sensitive civilian buildings are no longer restricted to specific theaters; if they exist in a theater, they are allowed. Military buildings, incuding the theater-sensitive Misile Silo and Pillbox in RA, are always usable, no matter whether the theater has graphics for them. * On Interior maps, the "ConvertRaObsoleteClear" logic will now generate spots of passable terrain outside the map border on any point where passable terrain touches the border, to allow potential reinforcements to enter the map there. * Removed all hardcoded special code per vehicle and per building, in preparation for exporting all editor settings to ini.

+++ CnCTDRAMapEditor/Interface/ITechnoType.cs
@@ -17,7 +17,7 @@ namespace MobiusEditor.Interface
     public interface ITechnoType: IBrowsableType
     {
         /// <summary>Object ID</summary>
-        sbyte ID { get; }
+        int ID { get; }
         /// <summary>Object ini name</summary>
         string Name { get; }
         /// <summary>True if this object has a weapon. This affects the default orders for placing it on the map.</summary>
